api/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
anyio==3.6.2
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dostoevsky==0.6.0
fastapi==0.89.1
fasttext==0.9.2
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
joblib==1.2.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.3
pydantic==1.10.4
python-multipart==0.0.6
razdel==0.5.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
termcolor==2.2.0
tflearn==0.5.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.3
wrapt==1.15.0
